#9755 S2S_Revision of the 'share requests' directory_ Part 1 - review fixes
