+++ sormas-api/src/main/resources/enum.properties
