+++ sormas-backend/src/main/java/de/symeda/sormas/backend/sormastosormas/entities/eventparticipant/EventParticipantShareDataBuilder.java
@@ -24,6 +24,7 @@ import de.symeda.sormas.api.event.EventParticipantDto;
 import de.symeda.sormas.api.sormastosormas.event.SormasToSormasEventParticipantDto;
 import de.symeda.sormas.api.sormastosormas.share.incoming.SormasToSormasEventParticipantPreview;
 import de.symeda.sormas.api.utils.ValidationRuntimeException;
+import de.symeda.sormas.backend.event.EventFacadeEjb;
 import de.symeda.sormas.backend.event.EventParticipant;
 import de.symeda.sormas.backend.event.EventParticipantFacadeEjb;
 import de.symeda.sormas.backend.sormastosormas.share.ShareDataBuilder;
@@ -74,6 +75,18 @@ public class EventParticipantShareDataBuilder
 	public SormasToSormasEventParticipantPreview doBuildShareDataPreview(EventParticipant eventParticipant, ShareRequestInfo requestInfo) {
 		Pseudonymizer pseudonymizer = dataBuilderHelper.createPseudonymizer(requestInfo);
 
-		return dataBuilderHelper.getEventParticipantPreview(eventParticipant, pseudonymizer);
+		return getEventParticipantPreview(eventParticipant, pseudonymizer);
+	}
+
+	public SormasToSormasEventParticipantPreview getEventParticipantPreview(EventParticipant eventParticipant, Pseudonymizer pseudonymizer) {
+		SormasToSormasEventParticipantPreview preview = new SormasToSormasEventParticipantPreview();
+
+		preview.setUuid(eventParticipant.getUuid());
+		preview.setPerson(dataBuilderHelper.getPersonPreview(eventParticipant.getPerson()));
+		preview.setEvent(EventFacadeEjb.toReferenceDto(eventParticipant.getEvent()));
+
+		pseudonymizer.pseudonymizeDto(SormasToSormasEventParticipantPreview.class, preview, false, null);
+
+		return preview;
 	}
 }
