+++ sormas-backend/src/main/resources/META-INF/persistence.xml
@@ -94,7 +94,7 @@
 		 <properties>
          	<property name="hibernate.dialect" value="de.symeda.sormas.backend.ExtendedPostgreSQL94Dialect"/>
          	<property name="hibernate.transaction.jta.platform" value="org.hibernate.service.jta.platform.internal.SunOneJtaPlatform"/>
-			<property name="hibernate.show_sql" value="false" />
+			<property name="hibernate.show_sql" value="true" />
 			<property name="hibernate.format_sql" value="false" />
          	<!-- don't validate on release, otherwise the automatic db update won't work
          	<property name="hibernate.hbm2ddl.auto" value="validate"/>
